gRPC_base/requirements.txt
grpcio==1.14.0
grpcio-tools==1.14.0